Tidy (FeedExportPrinter)
